Moving the doccumented parts of Source and Scene out of the dev folder and into the main code base
